fix 64 bit warnings

+++ src/SharedUtilities/Utils.cpp
@@ -462,7 +462,7 @@ namespace Utils
             throw DMExceptionWithErrorCode(status);
         }
 
-        status = RegSetValueEx(hKey, propName.c_str(), 0, REG_SZ, reinterpret_cast<const BYTE*>(propValue.c_str()), (propValue.size() + 1) * sizeof(propValue[0]));
+        status = RegSetValueEx(hKey, propName.c_str(), 0, REG_SZ, reinterpret_cast<const BYTE*>(propValue.c_str()), static_cast<DWORD>((propValue.size() + 1) * sizeof(propValue[0])));
         if (status != ERROR_SUCCESS) {
             RegCloseKey(hKey);
             throw DMExceptionWithErrorCode(status);
@@ -533,7 +533,7 @@ namespace Utils
         }
 
         vector<wchar_t> buffer(charCount);
-        charCount = ::GetEnvironmentVariable(variableName.c_str(), buffer.data(), buffer.size());
+        charCount = ::GetEnvironmentVariable(variableName.c_str(), buffer.data(), static_cast<DWORD>(buffer.size()));
         if (charCount == 0)
         {
             throw DMExceptionWithErrorCode(GetLastError());
@@ -547,7 +547,7 @@ namespace Utils
         UINT size = GetSystemDirectory(0, 0);
 
         vector<wchar_t> buffer(size);
-        if (size != GetSystemDirectory(buffer.data(), buffer.size()) + 1)
+        if (size != GetSystemDirectory(buffer.data(), static_cast<DWORD>(buffer.size()) + 1))
         {
             throw DMException("Error: failed to retrieve system folder.");
         }
@@ -684,7 +684,7 @@ namespace Utils
                 {
                     DWORD readByteCount = 0;
                     vector<char> readBuffer(bytesAvailable + 1);
-                    if (ReadFile(stdOutReadHandle.Get(), readBuffer.data(), readBuffer.size() - 1, &readByteCount, NULL) || readByteCount == 0)
+                    if (ReadFile(stdOutReadHandle.Get(), readBuffer.data(), static_cast<DWORD>(readBuffer.size() - 1), &readByteCount, NULL) || readByteCount == 0)
                     {
                         readBuffer[readByteCount] = '\0';
                         output += readBuffer.data();
@@ -733,13 +733,13 @@ namespace Utils
         TRACE(__FUNCTION__);
 
         DWORD destinationSize = 0;
-        if (!CryptBinaryToString(reinterpret_cast<unsigned char*>(buffer.data()), buffer.size(), CRYPT_STRING_BASE64, nullptr, &destinationSize))
+        if (!CryptBinaryToString(reinterpret_cast<unsigned char*>(buffer.data()), static_cast<DWORD>(buffer.size()), CRYPT_STRING_BASE64, nullptr, &destinationSize))
         {
             throw new DMException("Error: cannot obtain the required size to encode buffer into base64.");
         }
 
         vector<wchar_t> destinationBuffer(destinationSize);
-        if (!CryptBinaryToString(reinterpret_cast<unsigned char*>(buffer.data()), buffer.size(), CRYPT_STRING_BASE64, destinationBuffer.data(), &destinationSize))
+        if (!CryptBinaryToString(reinterpret_cast<unsigned char*>(buffer.data()), static_cast<DWORD>(buffer.size()), CRYPT_STRING_BASE64, destinationBuffer.data(), &destinationSize))
         {
             throw new DMException("Error: cannot convert binary stream to base64.");
         }
